portfolio-report/requirements.txt
requests>=2.32.0
python-dateutil>=2.9.0
PyYAML>=6.0
numpy-financial>=1.0.0
chinesecalendar>=1.9.0

